requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==3.2.16
django-crispy-forms==2.0
django-filter==23.3
django-login-required-middleware==0.9.0
django-widget-tweaks==1.5.0
pytz==2023.3.post1
sqlparse==0.4.4
